Use inspect module in add_movie for GratingMovie
Previously, if 'phase' was added as a parameter to add_movie it would
not be passed to GratingMovie.create_movie and would be ignored.
By using the inspect module, we can ensure that any arguments to
GratingMovie.__init__ and GratingMovie.create_movie are automatically
supported by FilterSimulator.add_movie.

We also created a test case, the setup of which can be reused to test
a variety of functionality of FilterSimulator.