Pointnet multiple runs (#410)
I've implemented the feature to run multiple PointNet simulations with a single config file.

The specifications and how to use is written in the manual page. The demo is written under the 'point450_glifs' directory. I did not make specific test cases for this feature as I did not find an appropriate section for this.